// frontend/generated/models/TransactionDto.ts
export interface TransactionDto {
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    transactionId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof TransactionDto
     */
    fromAccount?: AccountDto;
    /**
     * 
     * @type {AccountDto}
     * @memberof TransactionDto
     */
    toAccount?: AccountDto;
    /**
     * 
     * @type {MilestoneDto}
     * @memberof TransactionDto
     */
    milestone?: MilestoneDto;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    amount?: number;
    /**
     * 
     * @type {TransactionTypeDto}
     * @memberof TransactionDto
     */
    type?: TransactionTypeDto;
    /**
     * 
     * @type {TransactionStatusDto}
     * @memberof TransactionDto
     */
    status?: TransactionStatusDto;
    /**
     * 
     * @type {PaymentMethodDto}
     * @memberof TransactionDto
     */
    paymentMethod?: PaymentMethodDto;
    /**
     * 
     * @type {string}
     * @memberof TransactionDto
     */
    notes?: string;
    /**
     * 
     * @type {Date}
     * @memberof TransactionDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TransactionDto
     */
    updatedAt?: Date;
}