// frontend/generated/models/UpdateWithdrawDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateWithdrawDto
 */
export interface UpdateWithdrawDto {
    /**
     * 
     * @type {number}
     * @memberof UpdateWithdrawDto
     */
    transactionId?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateWithdrawDto
     */
    transactionStatus?: string;
}

/**
 * Check if a given object implements the UpdateWithdrawDto interface.
 */
export function instanceOfUpdateWithdrawDto(value: object): value is UpdateWithdrawDto {
    return true;
}

export function UpdateWithdrawDtoFromJSON(json: any): UpdateWithdrawDto {
    return UpdateWithdrawDtoFromJSONTyped(json, false);
}

export function UpdateWithdrawDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateWithdrawDto {
    if (json == null) {
        return json;
    }
    return {
        
        'transactionId': json['transactionId'] == null ? undefined : json['transactionId'],
        'transactionStatus': json['transactionStatus'] == null ? undefined : json['transactionStatus'],
    };
}

export function UpdateWithdrawDtoToJSON(json: any): UpdateWithdrawDto {
    return UpdateWithdrawDtoToJSONTyped(json, false);
}

export function UpdateWithdrawDtoToJSONTyped(value?: UpdateWithdrawDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'transactionId': value['transactionId'],
        'transactionStatus': value['transactionStatus'],
    };
}